A copy of the kcdc3 post-receive hook
As currently exists in /srv/git/kcdc3.git/hooks
and owned by the git user. This file is not
automatically updated if you change it (yet)
so plan on keeping it in sync if you make any
changes.

Currently, the post-receive hook has the actual
filecopy (cp -r) commented out while I investigate
moving this script to execute under the www-data
user.